requirements.txt
openai
python-dotenv
black
tqdm
ebooklib
beautifulsoup4
lxml
